star tracker orienations values are now int16

+++ src/boards/star_tracker/star_tracker_app.c
@@ -3,6 +3,7 @@
 #include "olm_file_cache.h"
 #include "star_tracker.h"
 #include "star_tracker_app.h"
+#include <stdint.h>
 
 void
 star_tracker_app_async(void *data, olm_file_cache_t *fread_cache) {
@@ -10,9 +11,9 @@ star_tracker_app_async(void *data, olm_file_cache_t *fread_cache) {
 
     if (star_tracker_coordinates(&coor) >= 0) {
         CO_LOCK_OD();
-        OD_orienation.rightAscension = coor.right_ascension;
-        OD_orienation.declination = coor.declination;
-        OD_orienation.roll = coor.roll;
+        OD_orienation.rightAscension = (int16_t)coor.right_ascension;
+        OD_orienation.declination = (int16_t)coor.declination;
+        OD_orienation.roll = (int16_t)coor.roll;
         OD_orienation.timestamp = 0; // TODO fix
         CO_UNLOCK_OD();
     } else {
